// src/enums/notification.ts
export enum NotificationTypeEnum {
  WHOLESALER_PRODUCT_REQUEST_CONNECT = "wholesaler_product_request_connect",
  PRODUCT_VERSION_PUBLISHED = "product_version_published",
  PRODUCT_PROCESS_COMPLETED = "product_process_completed",
  SUPPLIER_INVITE_ACCEPTED = "supplier_invite_accepted",
  WHOLESALER_PRODUCT_CONNECT_ACCEPTED = "wholesaler_product_connect_accepted",
  WHOLESALER_PRODUCT_CONNECT_REJECTED = "wholesaler_product_connect_rejected",
  DOCUMENT_SHARED = "document_shared",
  NLF_PROCESS_COMPLETE = "nlf_process_completed",
}