// src/enums/socket-connect.ts
export enum RoomEventEnum {
  JOIN_ROOM = "JOIN_ROOM",
  LEFT_ROOM = "LEFT_ROOM",
  ERROR = "ERROR",
}

export enum DisconnectEventEnum {
  SERVER_DISCONNECT = "io server disconnect",
  CLIENT_DISCONNECT = "io client disconnect",
}

export enum NotificationEnum {
  PRODUCT_EVENT = "PRODUCT_EVENT",
  COMPANY_EVENT = "COMPANY_EVENT",
  DOCUMENT_EVENT = "DOCUMENT_EVENT",
}